+++ prept/engine.py
@@ -24,6 +24,8 @@ class GenerationEngine:
 
     This class provides a rich interface for manipulating the generation
     time behavior.
+
+    .. versionadded:: 0.2.0
     """
     def __init__(self) -> None:
         self._file_processors: OrderedDict[str, list[ProcessorFunctionT]] = OrderedDict()