+++ docs/changelog.rst
@@ -11,16 +11,20 @@ v0.2.0
 
 - Add :attr:`~BoilerplateInfo.variable_input_mode` option to change behaviour of variables input prompt
 - Add support for :ref:`template paths <guide-templating--template-paths>`
+- Add support for :ref:`dynamic generation <guide-dynamic-generation>`
 
 **Enhancements and Changes**
 
 - :class:`PreptCLIError` now supports proper indentation formatting of multiline error message and hint
 - :meth:`TemplateProvider.render` has been renamed to :meth:`TemplateProvider.process_content`
 - :program:`prept info` now shows the basic details (name, required/optional, summary) of template variables
+- :attr:`~BoilerplateInfo.template_provider` now takes spec in standard Python module format i.e. ``module:object``
 
 **Fixes**
 
 - Fix :attr:`~BoilerplateInfo.allow_extra_variables` not having any effect
+- Fix template provider not resolving through class name when :func:`get_prept_template_provider` function was defined
+- Fix resolution failure for template providers from modules present in current working or boilerplate directory
 
 v0.1.0
 ~~~~~~